# openapi.yaml
swagger: "2.0"
info:
  description: "This application is part of a selective process at Dextra."

  version: "1.0.0"

  title: "Harry Potter Characters - Dextra Challenge"
host: "localhost:9090"
...
schemes:
  - "http"
paths:
...
  /users:
    post:
      tags:
        - "User"
      summary: "Create user"
      operationId: "addUser"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "user"
          description: "Created user object"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        201:
          description: "User created sucessfully"
          schema:
            type: "object"
            properties:
              message:
                type: "string"
                example: "Usuário criado com sucesso!"
        400:
          description: "Some error occurred"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  User:
    type: "object"
    properties:
      nome:
        type: "string"
      email:
        type: "string"
      senha:
        type: "string"
    xml:
      name: "User"
...